+++ CHANGELOG.md
@@ -1,4 +1,6 @@
 ## [Unreleased]
+  - 💄 [@mixte/snippets] 使 `Ant Design X` 的 `Bubble` 支持通过 `content` 传递 Vue 组件实现自定义渲染内容 *
+  - 💄 [@mixte/snippets] 使 `Ant Design X` 的 `BubbleList` 支持通过 `items[].content` 传递 Vue 组件实现自定义渲染内容 *
 
 ## [v3.4.0-beta.6]
   - 📅 2025-03-25
@@ -8,7 +10,7 @@
 
 ## [v3.4.0-beta.5]
   - 📅 2025-03-21
-  - 🌟 [@mixte/snippets] 为 `Ant Design X` 新增了一个 `defineBubbleListRoles` 工具方法, 用于定义 BubbleList 气泡默认属性 *
+  - 🌟 [@mixte/snippets] 为 `Ant Design X` 新增了一个 `defineBubbleListRoles` 工具方法, 用于定义 `BubbleList` 气泡默认属性 *
   - 🌟 [@mixte/snippets] 新增 `Ant Design X` 的 `SenderHeader` 组件 *
   - 🌟 [@mixte/snippets] 新增 `Ant Design X` 的 `XRequest` 和 `XStream` 导出 *
   - 💄 [@mixte/snippets] 使 `Ant Design X` 的 `Bubble` 支持通过 `messageRender` 传递 Vue 组件实现自定义渲染内容 *
@@ -24,7 +26,7 @@
 ## [v3.4.0-beta.3]
   - 📅 2025-03-20
   - 🌟 [mixte] 新增 `transformKeys` 函数, 用于转换传入对象的所有键名
-  - 🌟 [mixte] 新增 `assertPlainObject` 函数, 用于判断传入参数是否是纯粹的对象，并进行类型断言
+  - 🌟 [mixte] 新增 `assertPlainObject` 函数, 用于判断传入参数是否是纯粹的对象, 并进行类型断言
   - 🌟 [@mixte/snippets] 新增 `Ant Design X` 的 `Welcome` 组件 *
   - 🌟 [@mixte/snippets] 新增 `Ant Design X` 的 `XProvider` 组件 *
   - 🌟 [@mixte/snippets] 新增 `Ant Design X` 的 `Bubble` 和 `BubbleList` 组件 *
@@ -35,7 +37,7 @@
   - 🌟 [mixte] 新增 `pick` 和 `omit` 函数, 支持从对象中选择或排除指定属性
   - 🌟 [mixte] 新增 `deepFind` 函数, 用于在嵌套数据结构中深度查找满足条件的第一个元素
   - 🌟 [mixte] 新增 `deepSome` 函数, 用于在嵌套数据结构中深度检查是否存在满足条件的元素
-  - 🌟 [mixte] 新增 `deepForEach` 函数, 用于在嵌套结构数据中深度遍历所有元素，并为每个元素执行回调函数
+  - 🌟 [mixte] 新增 `deepForEach` 函数, 用于在嵌套结构数据中深度遍历所有元素, 并为每个元素执行回调函数
   - 💄 [@mixte/mel-components] 优化组件 `MelSelect` 的 `MelSelectOption` 类型
   - 💄 [@mixte/snippets] 完善 `Ant Design X` 的 `Sender` 组件 *
   - ⚠️ [@mixte/snippets] 修改 `Ant Design X` 的 `init-veaury` 更名为 `patch-for-react-19` *